[refactor] move metadata type defs
The new file path is easier to interpret at a glance, and was suggested
in <https://github.com/nextstrain/auspice/issues/2053>
